requirements.txt
celery==4.4.2
Django==3.0.4
Flask==1.1.1
jsonfield==3.1.0
kombu==4.6.8
marshmallow==3.5.1
numpy==1.18.2
pandas==1.0.3
pika==1.1.0
PyJWT>=1.4.0
pymongo==3.10.1
pymssql==2.1.4
pyodbc==4.0.30
python-dateutil==2.8.1
requests==2.23.0
SQLAlchemy==1.3.15
sshtunnel==0.1.5
textdistance==4.1.5
typing==3.7.4.1
zeep==3.4.0
PyYAML==5.3
datetime~=4.3
schedule~=0.6.0
python-json-logger
elasticsearch
CMRESHandler
flask-cors
Flask-JWT
flask-jwt-simple
flask_restful
facebook_business
pyhumps
